requirement.txt
requests
beautifulsoup4
pandas
SQLAlchemy
boto3
python-dotenv